requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
boto3==1.34.59
botocore==1.34.59
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
discord.py==2.3.2
frozenlist==1.4.1
google-auth==2.28.2
idna==3.6
importlib_metadata==7.0.2
jaraco.classes==3.3.1
jmespath==1.0.1
keyring==24.3.1
keyrings.google-artifactregistry-auth==1.1.2
more-itertools==10.2.0
multidict==6.0.5
pluggy==1.4.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.2
requests==2.31.0
rsa==4.9
s3transfer==0.10.0
six==1.16.0
urllib3==2.0.7
yarl==1.9.4
zipp==3.17.0
